mmio modified to use unsigned longs instead of ints

+++ utils/mmio.h
@@ -18,12 +18,12 @@ typedef char MM_typecode[4];
 char *mm_typecode_to_str(MM_typecode matcode);
 
 int mm_read_banner(FILE *f, MM_typecode *matcode);
-int mm_read_mtx_crd_size(FILE *f, int *M, int *N, int *nz);
-int mm_read_mtx_array_size(FILE *f, int *M, int *N);
+int mm_read_mtx_crd_size(FILE *f, unsigned long *M, unsigned long *N, unsigned long long *nz);
+int mm_read_mtx_array_size(FILE *f, unsigned long *M, unsigned long *N);
 
 int mm_write_banner(FILE *f, MM_typecode matcode);
-int mm_write_mtx_crd_size(FILE *f, int M, int N, int nz);
-int mm_write_mtx_array_size(FILE *f, int M, int N);
+int mm_write_mtx_crd_size(FILE *f, unsigned long M, unsigned long N, unsigned long long nz);
+int mm_write_mtx_array_size(FILE *f, unsigned long M, unsigned long N);
 
 
 /********************* MM_typecode query fucntions ***************************/
@@ -118,15 +118,15 @@ int mm_is_valid(MM_typecode matcode);		/* too complex for a macro */
 
 /*  high level routines */
 
-int mm_write_mtx_crd(char fname[], int M, int N, int nz, int I[], int J[],
+int mm_write_mtx_crd(char fname[], unsigned long M, unsigned long N, unsigned long long nz, unsigned long I[], unsigned long J[],
 		 double val[], MM_typecode matcode);
-int mm_read_mtx_crd_data(FILE *f, int M, int N, int nz, int I[], int J[],
+int mm_read_mtx_crd_data(FILE *f, unsigned long M, unsigned long N, unsigned long long nz, unsigned long I[], unsigned long J[],
 		double val[], MM_typecode matcode);
-int mm_read_mtx_crd_entry(FILE *f, int *I, int *J, double *real, double *img,
+int mm_read_mtx_crd_entry(FILE *f, unsigned long *I, unsigned long *J, double *real, double *img,
 			MM_typecode matcode);
 
-int mm_read_unsymmetric_sparse(const char *fname, int *M_, int *N_, int *nz_,
-                double **val_, int **I_, int **J_);
+int mm_read_unsymmetric_sparse(const char *fname, unsigned long *M_, unsigned long *N_, unsigned long long *nz_,
+                double **val_, unsigned long **I_, unsigned long **J_);
